+++ src/main/java/com/cfjst/piggy/dao/TeacherDao.java
@@ -17,7 +17,7 @@ public interface TeacherDao {
      * @param id 学号或工号
      */
     @Select("select * from teacher where id=#{id}")
-    public Teacher findById(Integer id);
+    public Teacher findById(Long id);
     
     /**
      * 添加用户
@@ -30,7 +30,7 @@ public interface TeacherDao {
      * @param id 学号或工号
      */
      @Delete("DELETE FROM `sbff`.`teacher` WHERE (`id` = #{id})")
-    public void delById(Integer id);
+    public void delById(Long id);
 
     /**
      * 修改用户信息
