+++ src/test/java/com/cfjst/piggy/dao/StudentDaoTests.java
@@ -42,7 +42,7 @@ public class StudentDaoTests {
     public void findByIdTest()  {
         //查找学生测试
 
-        Integer id = 2016051094;
+        Long id = Long.valueOf(2016051094);
         assertEquals("黑富福", dao.findById(id).getName());
 
     }
@@ -50,7 +50,7 @@ public class StudentDaoTests {
     @Test
     public void addTest()  {
         //添加、删除学生测试
-        Integer id = 1111111111;
+        Long id = Long.valueOf(1111111111) ;
         Student student = new Student(id,"123456","绿富福");        
         dao.add(student);        
         assertEquals("绿富福", dao.findById(id).getName());
@@ -63,7 +63,7 @@ public class StudentDaoTests {
     @Test
     public void updateTest()  {
         //修改学生信息测试
-        Integer id = 2016051094;
+        Long id = Long.valueOf(2016051094);
         Student student = new Student(id,"123456","白富福");        
         dao.update(student);        
         assertEquals("白富福", dao.findById(id).getName());
